pyproject.toml
[tool.poetry]
name = "flask-mongodb-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "flask_mongodb_test"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.2"
flask-mongoengine = "^1.0.0"
flask-restful = "^0.3.9"


[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
